Window prefix array
We keep track of known window prefix names and their column positions.  This array is saved to settings. Using the array We know used and free column positions.

A new button Close all prefixes is used to clear all windows with a prefix.

Help tips tell which prefix and column positions are known to the script.

Icon count is saved for every column position to keep track where AutoContinue  should put new icons.