Save relative paths to Json file
Fix paths in the Json file so that they are relative to the save directory if they have a common prefix.

Relative paths make it easier to move files and corresponding Json file around or even share it with someone else.

For example if save directory  is /Telescopes/TelescopeLive/NGC104 and file name is /Telescopes/TelescopeLive/NGC104/Lights/Red/NGC104_R.fits we save the file name as Lights/Red/NGC104_R.fits. If saveDir is a prefix of file name we save the full path. When Json file is read files are restored back with a full path.