Use file name in addition to view id to find images
In PixInsight 1.9.3 when loading an image from a file the view id is a generated name like Image32 instead of a file name like it used to be.  In xisf files saved with 1.9.3 the view id created originally is preserved when saving and loading. For compatibility with older data also the file name is checked to find for example Integration_R image for autocontinue. For non-xisf files that are loaded, view id is always a generated one. Maybe an option is needed for that too.